+++ cedar-policy-core/src/est/expr.rs
@@ -692,6 +692,23 @@ impl ExprBuilder for Builder {
         Expr::ExprNoExt(ExprNoExt::HasAttr(HasAttrRepr::Simple { left: expr, attr }))
     }
 
+    /// `left has attr1.attr2. ...` (with an Arc)
+    fn extended_has_attr_arc(self, expr: Arc<Self::Expr>, attrs: NonEmpty<SmolStr>) -> Self::Expr {
+        // Single attribute == a simple has
+        if attrs.tail.is_empty() {
+            Expr::ExprNoExt(ExprNoExt::HasAttr(HasAttrRepr::Simple {
+                left: expr,
+                attr: attrs.head,
+            }))
+        } else {
+            // Extended has form
+            Expr::ExprNoExt(ExprNoExt::HasAttr(HasAttrRepr::Extended {
+                left: expr,
+                attr: attrs,
+            }))
+        }
+    }
+
     /// `left like pattern`
     fn like(self, expr: Expr, pattern: ast::Pattern) -> Expr {
         Expr::ExprNoExt(ExprNoExt::Like {